# Conflicts:
#	src/backend/agentchat/api/services/chat.py
#	src/backend/agentchat/services/mcp_agent/agent.py

+++ src/backend/agentchat/services/rag_handler.py
@@ -24,18 +24,20 @@ class RagHandler:
 
     @classmethod
     async def mix_retrival_documents(cls, query_list, knowledges_id, search_field="summary"):
-        es_documents, milvus_documents = await MixRetrival.mix_retrival_documents(query_list, knowledges_id, search_field)
 
-        # 先对ES和Milvus结果分别排序
-        es_documents.sort(key=lambda x: x.score, reverse=True)
-        milvus_documents.sort(key=lambda x: x.score, reverse=True)
+        if app_settings.rag.enable_elasticsearch:
+            es_documents, milvus_documents = await MixRetrival.mix_retrival_documents(query_list, knowledges_id, search_field)
+            # 先对ES和Milvus结果分别排序
+            es_documents.sort(key=lambda x: x.score, reverse=True)
+            milvus_documents.sort(key=lambda x: x.score, reverse=True)
+            all_documents = es_documents + milvus_documents
+        else:
+            all_documents = await MixRetrival.retrival_milvus_documents(query_list, knowledges_id, search_field)
 
         # 合并并去重，保留分数更高的文档
         documents = []
         seen_chunk_ids = set()
-        
-        # 创建一个合并的文档列表
-        all_documents = es_documents + milvus_documents
+
         # 按分数从高到低排序
         all_documents.sort(key=lambda x: x.score, reverse=True)
         
@@ -89,19 +91,21 @@ class RagHandler:
 
 
     @classmethod
-    async def retrieve_ranked_documents(cls, query, collection_names, index_names, min_score: Optional[float]=None,
+    async def retrieve_ranked_documents(cls, query, collection_names, index_names=None, min_score: Optional[float]=None,
                         top_k: Optional[int]=None, needs_query_rewrite: bool=True):
         """
-            处理 RAG 流程：查询重写、文档检索、重排序、结果过滤和拼接。
+        处理 RAG 流程：查询重写、文档检索、重排序、结果过滤和拼接。
 
-            参数:
-                query (str): 用户查询。
-                knowledges_id (str): 知识库 ID。
-                min_score (float): 文档最低分数阈值，默认为配置中的值。
-                top_k (int): 召回文档的个数。
+        Args:
+            query (str): 用户查询。
+            collection_names (list[str]): 向量知识库 集合ID。
+            index_names (list[str]): ES关键词库 集合ID。
+            min_score (float): 文档最低分数阈值，默认为配置中的值。
+            top_k (int): 召回文档的个数。
+            needs_query_rewrite (bool): 是否需要开启Query重写，默认开启
 
-            返回:
-                str: 拼接后的最终结果。
+        Returns:
+            str: 拼接后的最终结果。
             """
         if min_score is None:
             min_score = app_settings.rag.retrival.get('min_score')
@@ -142,5 +146,6 @@ class RagHandler:
 
     @classmethod
     async def delete_documents_es_milvus(cls, file_id, knowledge_id):
-        await es_client.delete_documents(file_id, knowledge_id)
+        if app_settings.rag.enable_elasticsearch:
+            await es_client.delete_documents(file_id, knowledge_id)
         await milvus_client.delete_by_file_id(file_id, knowledge_id)
